Changed grow/shrink logic to avoid equality tests of floating point numbers.
To  compute ball's  new  radius, the  "current"  grow/shrink state  of
the  ball  is  determined  by  testing  a  couple  of  floating  point
numbers  for equality.  It  appears  that these  tests  can fail  even
when  comparing results  of  the same  computation, sometimes  causing
unexpected  grow/shrink behaviour.  It becomes  apparent with  certain
non-standard ball radius values as set  by the mapper, such as 0.1125,
but not with the default value of 0.25.

This  patch instead  keeps  track of  the current  state  by using  an
integer,  which  is  tested  for  positive/negative  values  and  zero
representing large, small and regular state, respectively.


git-svn-id: https://s.snth.net/svn/neverball/trunk@1226 78b8d119-cf0a-0410-b17c-f493084dd1d7