Implement a server/client-like game/replay architecture
The primary focus of this patch is splitting the ball/game module into
a "client" module and a "server" module.

The server module is responsible for running the game simulation,
while the client module mostly deals with drawing the game state to
the screen. The server communicates its state to the client through a
"proxy" module, which is simply a queue of "commands" encapsulating
the server state and events. The client reads the commands off the
proxy and runs them, updating its own state for display.  It can also
write them to a file, from which the same exact client state can then
be reconstructed by simply reading the commands back and adding them
to the proxy queue for the client to handle as described.

The primary purpose of the change was to eliminate game simulation
from the replay "operation chain", in order to eliminate most possible
causes for unreliable replays.  In ASCII art, the old scheme basically
looked like this:

    ...    -> input -> sim. -> display
    replay -> input -> sim. -> display

while the new one looks like this:

    ...    -> input -> server (sim.) -> proxy -> client -> display
    ...    -> ...   -> replay        -> proxy -> client -> display

In-detail description of the change:

    http://www.nevercorner.net/forum/viewtopic.php?pid=18842#p18842

git-svn-id: https://s.snth.net/svn/neverball/trunk@2714 78b8d119-cf0a-0410-b17c-f493084dd1d7