share/config: rewrite to reduce duplication and typing
The process of adding a new config option usually went something like
this: define a config symbol in one place, define a default value in
another, add code to init the current value with the default in
another, add code to write out the value in another, add code to read
the value in another.  It's a simple process, but it's a bit tedious
and naturally error-prone.

This patch reduces the process in the general case to: define a config
symbol in one place, associate the symbol with a name and a default
value in another, declare an external reference to the symbol in
another.  The special case is when integer options are represented in
non-numeric form in the config file (such as SDL key indices
represented by their human-readable names); this case still requires
the read/write routines also to be updated.

git-svn-id: https://s.snth.net/svn/neverball/trunk@2771 78b8d119-cf0a-0410-b17c-f493084dd1d7