Make sets.txt optional
It's not completely removed because there needs to be some way of
enforcing the order in which some sets are listed (e.g., Easy, then
Medium, then Hard).  Now the game first loads the sets listed in
sets.txt in the given order, then scans the root of the search path
for files starting with "set-" that it hasn't seen yet, and loads them
in alphabetic order.

This is aimed at making it easy to install new sets by simply dropping
a properly packaged ZIP archive in the user directory.

git-svn-id: https://s.snth.net/svn/neverball/trunk@2878 78b8d119-cf0a-0410-b17c-f493084dd1d7