Reorganize actions invoked by the F5-F12 keys
This brought to you by my inability to locate the screenshot key in
any fashion other than mashing keys and watching for a moment of
stutter to occur.

Here are the changes.

F5-F8 invoke cheat mode actions:

  F5  toggle look-around mode
  F6  toggle wireframe mode
  F8  take screenshots of all levels

F9-F12 invoke the remaining actions:

  F9  toggle FPS counter
  F10 toggle level pose mode / toggle HUD
  F12 take screenshot

The ability to toggle nice mode was removed.

git-svn-id: https://s.snth.net/svn/neverball/trunk@3807 78b8d119-cf0a-0410-b17c-f493084dd1d7